requirements.txt
openai>=1.57.0
streamlit>=1.37.0
python-dotenv>=1.0.1
chromadb>=0.5.5
sentence-transformers>=3.0.1
torch>=2.1.0
pydantic>=2.8.2
sqlalchemy>=2.0.30
tiktoken>=0.7.0
pypdf2>=3.0.1
python-docx>=1.1.2
